reafoctor : 예약 내역 페이지 전체적인 리팩터링
