+++ src/api/getMyActivity.ts
@@ -1,8 +1,12 @@
 import axiosInstance from '@/lib/axiosInstance';
+import { MyActivityType } from '@/types/myActivityPage';
 
 import { QueryFunctionContext } from '@tanstack/react-query';
 
-const getMyActivity = async ({ queryKey, pageParam }: QueryFunctionContext): Promise<any> => {
+const getMyActivity = async ({
+  queryKey,
+  pageParam,
+}: QueryFunctionContext): Promise<MyActivityType> => {
   const [, size] = queryKey;
   const cursorParam = pageParam ? `&cursorId=${pageParam}` : '';
   const dataSize = size ? `?size=${size}` : '';
