+++ src/components/reserveStatus/ReserveStatusContent.tsx
@@ -1,12 +1,12 @@
-import getMyActivity from '@/api/getMyActivity';
-import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { useState, useEffect } from 'react';
-import { useInView } from 'react-intersection-observer';
 import Calendar from 'react-calendar';
 import moment from 'moment';
 import getReservationYearAndMonth from '@/api/getReservationYearAndMonth';
 import queryKeys from '@/api/reactQuery/queryKeys';
 import { StyledReserveStatusCalendarWrapper } from '@/styles/StyledReserveStatusCalendar';
+import useInfiniteMyActivity from '@/hooks/useInfiniteMyActivity';
+import { Activity, ReservationData } from '@/types/reservationStatus';
 import ActivityDropDownBox from './ActivityDropDownBox';
 import ActivityDropDown from './ActivityDropDown';
 import PendingTileBlock from './PendingTileBlock';
@@ -14,36 +14,11 @@ import ConfimedTileBlock from './ConfirmedTileBlock';
 import CompletedTileBlock from './CompletedTileBlock';
 import ReservationModal from './reservationModal/ReservationModal';
 
-interface ReservationData {
-  date: string;
-  reservations: {
-    completed: number;
-    confirmed: number;
-    pending: number;
-  };
-}
-
-interface Activity {
-  id: number;
-  userId: number;
-  title: string;
-  description: string;
-  category: string;
-  price: number;
-  address: string;
-  bannerImageUrl: string;
-  rating: number;
-  reviewCount: number;
-  createdAt: string;
-  updatedAt: string;
-}
-
-interface ActivityData {
-  activities: Activity[];
-  totalCount: number;
-  cursorId: number;
-}
 const ReserveStatusContent = () => {
+  // 전체 체험 목록을 불러오는 custom hook
+  const { myActivityData } = useInfiniteMyActivity();
+  const activities = myActivityData?.pages.flatMap((page) => page.activities) || [];
+
   const currentDate = moment();
   // 달력에서 선택된 날짜를 저장하는 state
   const [selectedDate, setSelectedDate] = useState<{
@@ -64,14 +39,12 @@ const ReserveStatusContent = () => {
   // 선택된 activity state
   const [selectedActivity, setSelectedActivity] = useState<Activity | null>(null);
 
-  // useInfiniteQuery를 사용해 무한 스크롤 구현
-  const { data: activityData, fetchNextPage } = useInfiniteQuery<ActivityData>({
-    // queryKey: ['activity', 7],
-    queryKey: queryKeys.activities(),
-    queryFn: getMyActivity,
-    initialPageParam: 0,
-    getNextPageParam: (lastPage) => lastPage.cursorId,
-  });
+  useEffect(() => {
+    // activities가 존재할때 첫 번째 원소를 selectedActivity에 설정
+    if (activities.length > 0 && !selectedActivity) {
+      setSelectedActivity(activities[0]);
+    }
+  }, [activities, selectedActivity]);
 
   // 년도와 월을 이용해 예약 목록을 불러오는 함수
   const { data: reservationData } = useQuery({
@@ -82,15 +55,7 @@ const ReserveStatusContent = () => {
     ),
     queryFn: getReservationYearAndMonth,
   });
-  const activities = activityData?.pages.flatMap((page) => page.activities) || [];
-  useEffect(() => {
-    // activities가 존재할때 첫 번째 원소를 selectedActivity에 설정
-    if (activities.length > 0 && !selectedActivity) {
-      setSelectedActivity(activities[0]);
-    }
-  }, [activities, selectedActivity]);
 
-  const { inView, ref } = useInView();
   const getActiveMonth = (activeStartDate: moment.MomentInput) => {
     const changeDate = {
       year: moment(activeStartDate).format('YYYY'),
@@ -134,7 +99,9 @@ const ReserveStatusContent = () => {
 
   return (
     <div className="w-full min-w-[21.375rem] relative">
-      <h1 className="text-[32px] font-bold text-black mb-8">예약 현황</h1>
+      <h1 className="text-[32px] font-bold text-black mb-8 dark:text-darkMode-white-10">
+        예약 현황
+      </h1>
       {/* 드롭다운 박스에 선택된 activity title 표시 */}
       <ActivityDropDownBox
         selectedActivityTitle={selectedActivity ? selectedActivity?.title : ''}
@@ -143,12 +110,8 @@ const ReserveStatusContent = () => {
 
       {/* 전체 Activity 드롭다운으로 표시 */}
       {viewActivityDropDown && (
-        <div className="absolute w-[800px] ">
+        <div className="absolute w-[800px]  sm:w-full">
           <ActivityDropDown
-            activities={activities}
-            inView={inView}
-            fetchNextPage={fetchNextPage}
-            ref={ref}
             setViewActivityDropDown={setViewActivityDropDown}
             viewActivityDropDown={viewActivityDropDown}
             setSelectedActivity={setSelectedActivity}
@@ -167,14 +130,16 @@ const ReserveStatusContent = () => {
           tileContent={tileContent}
         />
       </StyledReserveStatusCalendarWrapper>
-      {viewReservationModal && selectedActivity && (
-        <ReservationModal
-          setViewReservationModal={setViewReservationModal}
-          selectedDate={selectedDate}
-          activitiyId={selectedActivity.id}
-          viewReservationModal={viewReservationModal}
-        />
-      )}
+      <div className="md:absolute md:top-[-10px] md:right-[-100px]">
+        {viewReservationModal && selectedActivity && (
+          <ReservationModal
+            setViewReservationModal={setViewReservationModal}
+            selectedDate={selectedDate}
+            activitiyId={selectedActivity.id}
+            viewReservationModal={viewReservationModal}
+          />
+        )}
+      </div>
     </div>
   );
 };
