+++ src/components/reserveStatus/reservationModal/ScheduleTimeDropDown.tsx
@@ -1,22 +1,5 @@
-import { useRef, useEffect, Dispatch, SetStateAction } from 'react';
-
-interface ScheduleType {
-  scheduleId: number;
-  startTime: string;
-  endTime: string;
-  count: {
-    declined: number;
-    confirmed: number;
-    pending: number;
-  };
-}
-
-interface ScheduleTimeDropDownProps {
-  schedule: ScheduleType[];
-  setSelelctedSchedule: React.Dispatch<React.SetStateAction<ScheduleType | null>>;
-  viewScheduleTimeDropDown: boolean;
-  setVieScheduleTimeDropDown: Dispatch<SetStateAction<boolean>>;
-}
+import { useRef, useEffect } from 'react';
+import { ScheduleType, ScheduleTimeDropDownProps } from '@/types/reservationStatus';
 
 const ScheduleTimeDropDown = ({
   schedule,
@@ -46,7 +29,7 @@ const ScheduleTimeDropDown = ({
   return (
     <div
       ref={dropdownRef}
-      className="w-full border border-gray-50 h-[170px] absolute z-50 bg-white rounded overflow-y-auto"
+      className="w-full border border-gray-50 h-[170px] absolute z-50 bg-white rounded overflow-y-auto dark:bg-darkMode-black-30 dark:text-white"
     >
       <ul>
         {schedule.map((item) => {
