+++ src/components/reserveStatus/reservationModal/ScheduleTimeDropDownBox.tsx
@@ -27,7 +27,7 @@ const ScheduleTimeDropDownBox = ({
       <input
         type="text"
         value={`${selectedSchedule?.startTime} ~ ${selectedSchedule?.endTime}`}
-        className="border border-gray-50 w-full rounded  pl-4 py-4 outline-none"
+        className="border border-gray-50 w-full rounded  pl-4 py-4 outline-none dark:bg-darkMode-black-30 dark:text-white"
       />
       <button type="button" className="cursor-pointer" onClick={handleClickDropDown}>
         <img
