+++ src/hooks/useSignup.ts
@@ -31,7 +31,7 @@ const useSignup = () => {
       });
 
       // 회원가입 성공후 알림, 로그인 페이지로 이동
-      navigate('/');
+      navigate('/login');
       toast.success('회원가입에 성공 했습니다!', {
         autoClose: 2000,
       });
